pyproject.toml
[tool.poetry]
name = "polyakov-mikhail-course-work-4"
version = "0.1.0"
description = ""
authors = ["Mikhail Polyakov <timadgi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
